[not verified] Docker: Make SFTP key auth easy, and adding plugins/themes less likely to break Backup/Scan (#17038)
* Docker: Make adding SFTP public key auth easy

The docker image being used for providing SFTP access, atmoz/sftp,
supports key-based authentication out of the box if you map public key
files to a particular directory in the container.

Set up a host directory mapping for that by default so all someone has
to do to make use of it is drop the public key file in the right place.

* Docker: Introduce compose-volumes.yml for custom plugins/themes

Our docker setup runs web service and SFTP out of two different
containers, as seems to be the accepted practice for docker.

Unfortunately, docker and docker-compose aren't really designed anymore
for use cases using bind mounts instead of docker-managed "volumes". The
only way to get the same set of bind mounts into two containers is to
declare them twice.

So it's easy for them to get out of sync, as was done by PR #16376 and
would probably happen if anyone uses the compose-extras.yml added in
PR #15824 to add plugins or themes. And if not, it would be onerous on
everyone doing so to have to repeat the stuff twice!

What's worse, we can't actually use the same set of bind mounts: the
ones for SFTP need to be under /home/wordpress/var/www/html, while for
the web service they need to be under /var/www/html. And we can't just
use the "fix" from https://github.com/atmoz/sftp#bindmount-dirs-from-another-location
as that requires the not-very-secure CAP_SYS_ADMIN (and, apparently,
some other stuff too that's not yet documented).

So this is the workaround:

* Declare the volumes making up the WordPress install in a separate
  file, docker/compose-volumes.yml.
* Use a simple shell script to convert that into a full docker-compose
  file with the needed bind mounts repeated for both services.

I didn't remove docker/compose-extras.yml, since there are still
possible use cases for it other than adding WordPress
plugins/themes/etc.